Agregados criterios: 1- De conjuntos con nota mas alta (con bug). 2- De conjuntos con promedio (solo lo base)

+++ Project-02-CorrectorExamenes/src/main/java/utn/disenio/criterio/Criterio.java
@@ -2,5 +2,5 @@ package utn.disenio.criterio;
 
 public interface Criterio {
 	
-	public Integer calcularNota(Integer pesoAlumno);
+	public Double calcularNota(Integer pesoAlumno);
 }
